Various improvements during move to GenDev env
Modifications
-------------
Upgraded to Amazon Linux 2 as the base image for the Linux build agent.

Changed the Linux build agent to automatically pick the latest AMI from
the correct category.

Upgraded to docker-compose 1.23.1 on the Linux build agent.

Build Linux image with ENA enabled so we can use m5 instances.

Create encrypted AMIs so that our build agents run on an encrypted
volume for added security.