End paths with semi-colons(ADV-17308)
Motivation
------------
Via further testing I noticed some inconsistencies with setx. If you don't end a path with a semi-colon, it adds a double quote at the end of the path, with a semi-colon, which causes issues with further calls to setx where you use the original path. To alleivate this, adding a semi-colon to the end prevents a double quote from added to the end.

Modification
-------------
 - Added semi-colons to end of paths